Implement ? operator, Result type codegen, binary expressions
- Result[T, E] return type -> Go (T, error)
- let x = expr? -> Go if err != nil pattern
- Ok(x) -> return x, nil
- Error(e) -> return zero, e
- Binary operators (+, -, *, /, ==, !=, <, >, etc.)
- Int maps to Go int (was int64)
- Ok/Error/Some/None as built-in constructors in checker

+++ testdata/sum_type.go
@@ -11,7 +11,7 @@ func (ApiResponseSuccess) isApiResponse() {}
 
 type ApiResponseErrorResponse struct {
 	Message string
-	Code int64
+	Code int
 }
 func (ApiResponseErrorResponse) isApiResponse() {}
